refactor: cell value formatting in project columns
